Добавляет функционал в презентер Убирает лишнее из презентера
