+++ .gitignore
@@ -14,5 +14,4 @@ node_modules/
 # Собранные файлы (допишите самостоятельно)
 public/bundle.js
 public/bundle.js.map
-src/main-copy.js
