+++ src/presenter/Movie.js
@@ -1,0 +1,1 @@
+import FilmCard from "../view/film-card";
